+++ src/utils.h
@@ -3,6 +3,8 @@
 #ifndef _IPTSD_SYSCALL_H_
 #define _IPTSD_SYSCALL_H_
 
+#include <stdint.h>
+
 #define iptsd_err(ERRNO, ARGS...) \
 	iptsd_utils_err(ERRNO, __FILE__, __LINE__, ##ARGS)
 
@@ -13,7 +15,7 @@ int iptsd_utils_write(int fd, void *buf, size_t count);
 int iptsd_utils_ioctl(int fd, unsigned long request, void *data);
 void iptsd_utils_err(int err, const char *file,
 	int line, const char *format, ...);
-unsigned iptsd_utils_msec_timestamp(void);
+uint64_t iptsd_utils_msec_timestamp(void);
 
 #endif /* _IPTSD_SYSCALL_H_ */
